--- src/components/Onboarding/Onboarding.css
@@ -1,30 +1,0 @@
-.root {
-  margin-left: 40%;
-  margin-top: 5%
-}
-.text-fields {
-  width: 40%;
-}
-.large-text {
-  font-size: large;
-  width: 40%;
-  font-weight: bold;
-  color: blue;
-}
-
-.error-msg {
-  color: red;
-}
-
-.button {
-  border: none;
-  color: white;
-  text-align: center;
-  text-decoration: none;
-  display: inline-block;
-  font-size: 16px;
-  cursor: pointer;
-}
-
-.button1 {background-color: #4CAF50;} /* Green */
-.button2 {background-color: #008CBA;} /* Blue */
