--- src/components/Onboarding/index.js
@@ -1,3 +1,0 @@
-import OnboardingScreen from './OnboardingScreen';
-
-export default OnboardingScreen;
