--- src/components/SignUp/SignUp.css
@@ -1,15 +1,0 @@
-.root {
-  position: relative;
-  min-height: 100vh;
-  margin-left: 40%;
-  margin-top: 5%;
-  position: relative;
-  min-height: 100vh;
-}
-.text-fields {
-  width: 40%;
-}
-
-.error-msg {
-  color: red;
-}
